+++ db/seed_bucket.js
@@ -1,11 +1,10 @@
-// Use the JS library to create a bucket.
+const prisma = require("./prisma");
 const { decode } = require("base64-arraybuffer");
 const supabase = require("./supabase");
-const getFileExtension = require("../utils/getFileExtension");
 const getFilesFromPath = require("../utils/getFilesFromPath");
-const allowedMimeTypes = require("../validators/allowedMimeTypes/allowedMimeTypes");
+const generateRandomIndex = require("../utils/generateRandomIndex");
 
-const seedBucket = async (bucketID = "drives") => {
+const seedBucket = async (userID, folders, bucketID = "drives", filesPath) => {
   const { data: bucketExists, error: bucketExistsError } =
     await supabase.storage.getBucket(bucketID);
 
@@ -22,27 +21,41 @@ const seedBucket = async (bucketID = "drives") => {
     await supabase.storage.createBucket(bucketID, { fileSizeLimit: "5MB" });
   }
 
+  const files = await getFilesFromPath(filesPath);
+  console.log("folders:", folders);
   // Upload files from uploadsPath
-
-  const files = await getFilesFromPath();
-
   for (const file of files) {
-    const { name: filename, parentPath } = file;
-    const ext = getFileExtension(filename);
-    const mimetype = allowedMimeTypes[ext];
-    const fileFullPath = `${parentPath}/${filename}`;
-    const buffer = await fsPromises.readFile(fileFullPath);
+    const randomFolderIndex = generateRandomIndex(files, 1);
+    const { id: folderID } = folders[randomFolderIndex] ?? {};
+    const { originalname, buffer, mimetype } = file;
     const fileBase64 = decode(buffer.toString("base64"));
-    // await supabase.storage
-    //   .from(bucketID)
-    //   .upload(`/${file.name}`, fileBase64, { contentType: mimetype });
+    const storagePath = `/${userID}${
+      folderID ? `/${folderID}` : ""
+    }/${originalname}`;
+
+    console.log("file:", file);
+    console.log("storagePath:", storagePath);
+    console.log("folderID:", folderID);
+    await supabase.storage
+      .from(bucketID)
+      .upload(storagePath, fileBase64, { contentType: mimetype });
+
+    await prisma.file.create({
+      data: {
+        name: originalname,
+        size: file.size,
+        url: storagePath,
+        accountId: userID,
+        folderId: folderID,
+      },
+    });
   }
 
-  console.log("buckets have been seeded!");
+  console.log("The bucket have been seeded!");
 };
 
 const emptyBucket = async (bucketID) => {
   await supabase.storage.emptyBucket(bucketID);
 };
 
-module.exports = { seedBucket };
+module.exports = seedBucket;